EmojiMaze/src/main/java/com/group22/world/worldLoader.java
package com.group22.world;

import java.io.BufferedReader;
import java.io.FileReader;
import java.io.IOException;

// loads file into string
public class worldLoader {
    public static String loadFileAsString(String path) {
        StringBuilder builder = new StringBuilder();

        try {
            BufferedReader br = new BufferedReader(new FileReader(path));
            String line;
            while ((line = br.readLine()) != null) {
                builder.append(line + "\n");
            }

            br.close();
        } catch (IOException e) {
            e.printStackTrace();
        }

        return builder.toString();
    }
}
